Added enclave pool inside the enclave manager and moved the creation logic inside the enclave creator
