+++ engine/server/engine/server/engine_server_service.go
@@ -19,9 +19,6 @@ type EngineServerService struct {
 
 	enclaveManager *enclave_manager.EnclaveManager
 
-	//TODO should check if we are going to leave it here or we move it inside the EnclaveManager
-	enclavePool *enclave_manager.EnclavePool
-
 	//The protected user ID for metrics analytics purpose
 	metricsUserID string
 
@@ -35,7 +32,6 @@ type EngineServerService struct {
 func NewEngineServerService(
 	imageVersionTag string,
 	enclaveManager *enclave_manager.EnclaveManager,
-	enclavePool *enclave_manager.EnclavePool,
 	metricsUserId string,
 	didUserAcceptSendingMetrics bool,
 	logsDatabaseClient centralized_logs.LogsDatabaseClient,
@@ -43,7 +39,6 @@ func NewEngineServerService(
 	service := &EngineServerService{
 		imageVersionTag:             imageVersionTag,
 		enclaveManager:              enclaveManager,
-		enclavePool:                 enclavePool,
 		metricsUserID:               metricsUserId,
 		didUserAcceptSendingMetrics: didUserAcceptSendingMetrics,
 		logsDatabaseClient:          logsDatabaseClient,
@@ -65,17 +60,7 @@ func (service *EngineServerService) CreateEnclave(ctx context.Context, args *kur
 		return nil, stacktrace.Propagate(err, "An error occurred parsing the log level string '%v':", args.ApiContainerLogLevel)
 	}
 
-	response := &kurtosis_engine_rpc_api_bindings.CreateEnclaveResponse{}
-
-	// TODO move all this logic inside the enclave manager
-	enclaveInfoFromPool, err := service.enclavePool.GetEnclave(args.EnclaveName)
-	if err == nil {
-		response.EnclaveInfo = enclaveInfoFromPool
-		return response, nil
-	}
-	logrus.Debugf("An error occurred when trying to get an enclave from the enclave pool")
-
-	newEnclaveInfo, err := service.enclaveManager.CreateEnclave(
+	enclaveInfo, err := service.enclaveManager.CreateEnclave(
 		ctx,
 		args.ApiContainerVersionTag,
 		apiContainerLogLevel,
@@ -87,7 +72,10 @@ func (service *EngineServerService) CreateEnclave(ctx context.Context, args *kur
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating new enclave with name '%v'", args.EnclaveName)
 	}
-	response.EnclaveInfo = newEnclaveInfo
+
+	response := &kurtosis_engine_rpc_api_bindings.CreateEnclaveResponse{
+		EnclaveInfo: enclaveInfo,
+	}
 
 	return response, nil
 }
